Use double splats to distinguish keyword arguments
This resolves a number of cases where Ruby would raise an exception such
as: `ArgumentError: wrong number of arguments (given 2, expected 1)` where
it identified a hash as a second argument and not keyword arguments.

These are cases when a hash was used for keyword arguments, in Ruby 3 a
double splat needs to be used to convert these hashes into keyword
arguments.